refactor(ui): lift useDelayedFlag/createDelayedFlag into @maka/ui
The side-conversation panel needed the main chat's rising-edge running-status
timing (useDelayedFlag + the 200ms delay), but the workbar feature layer cannot
import the renderer-legacy use-delayed-flag.ts / model-wait-state.ts (the
architecture gate walls features off from legacy), so #4655 added a feature-local
copy of useDelayedFlag and a `const RUNNING_STATUS_DELAY_MS = 200` to
quote-companion-panel.tsx — two authorities for one delay, a literal that would
drift, and no test coverage for the timing.

Move the pure createDelayedFlag (+ DelayedFlag, DelayedFlagScheduler), the
useDelayedFlag React binding, and RUNNING_STATUS_DELAY_MS into a new @maka/ui
`delayed-flag` module. The main chat (use-shell-live-turn) and the side chat
(quote-companion-panel) now import one implementation; the feature-local copy
and the renderer-legacy use-delayed-flag.ts are removed. Adds fake-timer unit
tests for createDelayedFlag (there were none), and drops the use-delayed-flag.ts
closure from the renderer-architecture ledger (net debt reduction).

Fixes #4844

Generated-by: Claude Code
