+++ packages/ui/src/index.ts
@@ -21,6 +21,7 @@ export * from './artifact-preview-registry.js';
 export * from './assistant-stream.js';
 export * from './chat-empty-hero.js';
 export * from './chat-model-helpers.js';
+export * from './delayed-flag.js';
 export * from './use-mounted-ref.js';
 export * from './session-setting-intent.js';
 export * from './components.js';
